File Structure For Weather Query
