+++ docs/sdk/code-reference/query/weather/latest.md
@@ -1,2 +1,2 @@
 # Weather Latest Function
-::: src.sdk.python.rtdip_sdk.queries.time_series.weather.latest
+::: src.sdk.python.rtdip_sdk.queries.weather.latest
