+++ docs/sdk/code-reference/query/weather/raw.md
@@ -1,2 +1,2 @@
 # Weather Raw Function
-::: src.sdk.python.rtdip_sdk.queries.time_series.weather.raw
+::: src.sdk.python.rtdip_sdk.queries.weather.raw
