+++ docs/sdk/code-reference/query/weather/weather_query_builder.md
@@ -1,2 +1,2 @@
 # Weather Query Builder
-::: src.sdk.python.rtdip_sdk.queries.weather_query_builder
+::: src.sdk.python.rtdip_sdk.queries.weather.weather_query_builder
