+++ src/sdk/python/rtdip_sdk/queries/weather/__init__.py
