+++ tests/sdk/python/rtdip_sdk/queries/weather/test_latest.py
@@ -24,10 +24,10 @@ from tests.sdk.python.rtdip_sdk.connectors.odbc.test_db_sql_connector import (
     MockedCursor,
 )
 from src.sdk.python.rtdip_sdk.connectors import DatabricksSQLConnection
-from src.sdk.python.rtdip_sdk.queries.time_series.weather.latest import (
+from src.sdk.python.rtdip_sdk.queries.weather.latest import (
     get_grid as latest_grid,
 )
-from src.sdk.python.rtdip_sdk.queries.time_series.weather.latest import (
+from src.sdk.python.rtdip_sdk.queries.weather.latest import (
     get_point as latest_point,
 )
